requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
mcp>=1.9.0
httpx>=0.28.0
PyJWT>=2.10.0
pydantic>=2.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
